Refactor package structure to follow modern Python best practices
Move core implementations from __init__.py to dedicated core.py module.

## Changes
- Create wlgen/core.py with all wordlist generation implementations
- Update wlgen/__init__.py to be a clean public API with imports
- Add __all__ and __version__ to __init__.py
- Update documentation references (CLAUDE.md, OPTIMIZATION_ANALYSIS.md)

## Benefits
- Cleaner separation of concerns
- Follows modern Python packaging conventions
- Makes codebase more maintainable
- All tests and benchmarks continue to work unchanged

## Backward Compatibility
- ✅ All public API remains unchanged
- ✅ Imports work identically (from wlgen import ...)
- ✅ All tests pass
- ✅ Benchmarks work correctly
